// src/app/modules/core/dmo/model/dmo.ts
export class Nwma {
  totalOutstandingNwma: string;
  noOfDaysInNwma: string;
  interestRate: string;
  interestPayable: string;
  interestPaid: string;
  perOfNwmaLimitUtilized: string;
}